Move insight services into Insights feature folders #388
