+++ code/FinanceManager.Application/Insights/Registration.cs
@@ -1,0 +1,20 @@
+using FinanceManager.Application.Insights.Diversification;
+using FinanceManager.Application.Insights.Generation;
+using FinanceManager.Application.Insights.Seeders;
+using FinanceManager.Domain.Services;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace FinanceManager.Application.Insights;
+
+internal static class Registration
+{
+    public static IServiceCollection AddInsightsApplication(this IServiceCollection services)
+    {
+        services.AddScoped<IFinancialInsightsAiGenerator, FinancialInsightsAiGenerator>()
+                .AddScoped<IDiversificationService, DiversificationService>()
+                .AddScoped<FinancialInsightsSeeder>();
+        // .AddScoped<ISeeder, FinancialInsightsSeeder>()
+
+        return services;
+    }
+}
