+++ code/FinanceManager.Application/Services/Seeders/GuestAccountSeeder.cs
@@ -1,6 +1,7 @@
 using FinanceManager.Application.FinancialAccounts.Bond.Seeders;
 using FinanceManager.Application.FinancialAccounts.Currencies.Seeders;
 using FinanceManager.Application.FinancialAccounts.Stock.Seeders;
+using FinanceManager.Application.Insights.Seeders;
 using FinanceManager.Application.Labels.Seeders;
 using FinanceManager.Domain.Enums;
 using FinanceManager.Domain.Repositories;
