+++ code/FinanceManager.Infrastructure/ServiceCollectionExtension.cs
@@ -1,12 +1,12 @@
 ﻿using FinanceManager.Application.FinancialAccounts.Stock.Market;
 using FinanceManager.Application.FinancialAccounts.Stock.Pricing;
+using FinanceManager.Application.Insights.Generation;
 using FinanceManager.Application.Labels.Setter;
 using FinanceManager.Application.Labels.Suggestions;
 using FinanceManager.Application.Options;
 using FinanceManager.Application.Services;
 using FinanceManager.Application.Services.Ai;
 using FinanceManager.Application.Services.ExternalServices;
-using FinanceManager.Application.Services.FinancialInsights;
 using FinanceManager.Domain.Entities.Bonds;
 using FinanceManager.Domain.Entities.FinancialAccounts.Currencies;
 using FinanceManager.Domain.Entities.Stocks;
