+++ code/FinanceManager.Tests.Unit/Api/Controllers/FinancialInsightsControllerTests.cs
@@ -1,5 +1,5 @@
 using FinanceManager.Api.Controllers;
-using FinanceManager.Application.Services.FinancialInsights;
+using FinanceManager.Application.Insights.Generation;
 using FinanceManager.Domain.Entities.Users;
 using FinanceManager.Domain.Repositories;
 using Microsoft.AspNetCore.Http;
