+++ code/FinanceManager.Tests.Unit/Application/Services/DiversificationServiceTests.cs
@@ -1,3 +1,4 @@
+using FinanceManager.Application.Insights.Diversification;
 using FinanceManager.Application.Services;
 using FinanceManager.Domain.Entities.Bonds;
 using FinanceManager.Domain.Entities.Currencies;
